+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add system/core/database.
 - Add system/core/session.
 - Add system/core/http/request.
+- Add system/core/http/response.
 
 ### Changed
 - Reorganize folder structure: separate Shift and public files.